update telecsope monitoring

+++ database/migrations/2025_05_18_200011_create_classroom_session_table.php
@@ -8,17 +8,19 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up(): void
     {
-        Schema::create('classroom_session', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('classroom_id');
-            $table->unsignedBigInteger('quiz_session_id');
-            $table->timestamps();
+        if (!Schema::hasTable('classroom_session')) {
+            Schema::create('classroom_session', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('classroom_id');
+                $table->unsignedBigInteger('quiz_session_id');
+                $table->timestamps();
 
-            $table->foreign('classroom_id')->references('id')->on('classrooms')->onDelete('cascade');
-            $table->foreign('quiz_session_id')->references('id')->on('quiz_sessions')->onDelete('cascade');
+                $table->foreign('classroom_id')->references('id')->on('classrooms')->onDelete('cascade');
+                $table->foreign('quiz_session_id')->references('id')->on('quiz_sessions')->onDelete('cascade');
 
-            $table->unique(['classroom_id', 'quiz_session_id']);
-        });
+                $table->unique(['classroom_id', 'quiz_session_id']);
+            });
+        }
     }
 
     public function down(): void
